# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tracks api",
        "description": "Управление сущностями проекта трекера произведений.",
        "contact": {
            "email": "xenx.efy@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/async",
            "description": "Main server"
        }
    ],
    "paths": {
...
        "/tracks": {
            "get": {
                "tags": [
                    "Composition"
                ],
                "summary": "Получение треков с возможностью сортировки.",
                "operationId": "getCompositions",
                "parameters": [
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "Сортироует треки по орпеделенному полю.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "title",
                                "last_played"
                            ]
                        }
                    },
                    {
                        "name": "sortType",
                        "in": "query",
                        "description": "Прямая или обратная сортировка.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Фильтрация по названию трека или фамилии композитора.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tags[]",
                        "in": "query",
                        "description": "Фильтрация по тегам",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                }
            },
            "post": {
                "tags": [
                    "Composition"
                ],
                "summary": "Создает композицию с информацией о ней.",
                "operationId": "storeCompositions",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "composerId",
                                    "tags"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Название произведения",
                                        "type": "string",
                                        "example": "Adazio"
                                    },
                                    "composerId": {
                                        "description": "Id композитора произведения",
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "tags": {
                                        "description": "Теги произведения",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response"
                    },
                    "201": {
                        "description": "Success creation"
                    }
                }
            }
        },
        "/tracks/{id}": {
            "put": {
                "tags": [
                    "Composition"
                ],
                "summary": "Обновляет информацию о треке.",
                "operationId": "updateComposition",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id обновляемого трека",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Формат входящих данных",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Название произведения",
                                        "type": "string",
                                        "example": "Adazio"
                                    },
                                    "composerId": {
                                        "description": "Id композитора произведения",
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "tags": {
                                        "description": "Теги произведения",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Composition"
                ],
                "summary": "Удалить трек.",
                "operationId": "deleteCompotition",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id удаляемого трека",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}